requirement.txt
python-dotenv==1.0.0
python-telegram-bot==20.0
google-api-python-client==2.78.0
google-auth==2.21.0
google-auth-oauthlib==0.8.0
google-auth-httplib2==0.1.0
gspread==5.7.0
oauth2client==4.1.3
